ml_project/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.12.12
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cffi==1.15.1
click==8.1.3
coverage==6.5.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
flake8==5.0.4
idna==3.4
ipykernel==6.16.1
ipython==8.5.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-dataclass==8.5.9
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mypy-extensions==0.4.3
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
pylint==2.15.5
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
svgwrite==1.4.3
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.5.0
typing-inspect==0.8.0
typing_extensions==4.4.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
wrapt==1.14.1
antlr4-python3-runtime==4.9.3
certifi==2022.9.24
charset-normalizer==2.1.1
hydra-core==1.2.0
idna==3.4
omegaconf==2.2.3
packaging==21.3
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12